broker: Delete instances on bind
A supervisor bug prevents pods from successfully completing a leader
election. However, deleting them all at once fixes the problem. This
bug shows up when applying a bind with multiple instances. In this
commit, we wait for all the pods to be updated after the bind
operation is applied by the user and then delete all the running pods
at once and we let the StatefulSet controller take over from there to
restore the desired state in the cluster.

More about the bug here: https://github.com/habitat-sh/habitat/issues/5264

Signed-off-by: Indradhanush Gupta <indra@kinvolk.io>